backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
httpx==0.27.0
faster-whisper==1.0.3
edge-tts==6.1.12
numpy==1.26.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0
python-multipart==0.0.9
